requirements.txt
absl-py==2.1.0
accelerate==0.18.0
aiohttp==3.9.3
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
brotli==1.0.9
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==2.0.4
comm==0.2.2
controlnet-aux==0.0.7
debugpy==1.6.7
decorator==5.1.1
diffusers==0.26.1
einops==0.7.0
entrypoints==0.4
executing==2.0.1
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.10.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
grpcio==1.62.1
huggingface-hub==0.22.1
idna==3.4
imageio==2.34.0
importlib-metadata==7.1.0
intel-cmplr-lib-ur==2025.0.3
intel-openmp==2025.0.3
ipykernel==6.29.3
ipython==8.12.0
jedi==0.19.1
jinja2==3.1.4
jupyter-client==7.3.4
jupyter-core==5.7.2
lazy-loader==0.3
lightning-utilities==0.11.1
markdown==3.6
markupsafe==3.0.2
matplotlib-inline==0.1.6
mkl==2025.0.1
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.5
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
opencv-python==4.9.0.80
packaging==24.0
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pip==23.3.1
platformdirs==4.2.0
prompt-toolkit==3.0.42
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pudb==2024.1.3
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydeprecate==0.3.2
pygments==2.17.2
pysocks==1.7.1
python-dateutil==2.9.0
pytorch-lightning==1.6.5
pywavelets==1.4.1
pyyaml==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.2
scikit-image==0.21.0
scipy==1.10.1
setuptools==68.2.2
six==1.16.0
stack-data==0.6.2
sympy==1.13.1
tbb==2022.0.0
tcmlib==1.2.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tifffile==2023.7.10
timm==0.6.7
tokenizers==0.13.3
torch==2.5.1
torchmetrics==1.3.2
torchvision==0.20.1
tornado==6.1
tqdm==4.66.2
traitlets==5.14.2
transformers==4.26.1
triton==3.1.0
typing-extensions==4.12.2
umf==0.9.1
urllib3==2.1.0
urwid==2.6.16
urwid-readline==0.15.1
wcwidth==0.2.13
werkzeug==3.0.1
wheel==0.41.2
yarl==1.9.4
zipp==3.18.1
